comprehensive_dev_server/database_explorer.py
import psycopg2
import pymysql
import sqlite3
from pymongo import MongoClient
from sqlalchemy import create_engine, inspect
from plantuml import PlantUML

# A global variable to store the database engine
db_engine = None

async def connect_database(connection_string: str, db_type: str):
    """Establish database connection."""
    global db_engine
    try:
        # For simplicity, we'll assume the connection string is a valid SQLAlchemy URI
        db_engine = create_engine(connection_string)
        # Test the connection
        with db_engine.connect() as connection:
            return {"status": f"{db_type} connection successful."}
    except Exception as e:
        db_engine = None
        return {"error": f"Connection failed: {e}"}

async def explore_schema(database_name: str):
    """Return complete schema information."""
    if not db_engine:
        return {"error": "Database not connected. Please connect to a database first."}
    
    try:
        inspector = inspect(db_engine)
        schema = {}
        for table_name in inspector.get_table_names():
            schema[table_name] = {
                "columns": [],
                "foreign_keys": [],
                "indexes": [],
            }
            for column in inspector.get_columns(table_name):
                schema[table_name]["columns"].append(column)
            for fk in inspector.get_foreign_keys(table_name):
                schema[table_name]["foreign_keys"].append(fk)
            for index in inspector.get_indexes(table_name):
                schema[table_name]["indexes"].append(index)
        return schema
    except Exception as e:
        return {"error": f"Schema exploration failed: {e}"}


async def generate_erd(database_name: str, erd_format: str):
    """Generate ERD diagrams."""
    if not db_engine:
        return {"error": "Database not connected. Please connect to a database first."}

    try:
        inspector = inspect(db_engine)
        plantuml_code = "@startuml\n"
        for table_name in inspector.get_table_names():
            plantuml_code += f"entity {table_name} {{\n"
            for column in inspector.get_columns(table_name):
                plantuml_code += f"  {column['name']} : {column['type']}\n"
            plantuml_code += "}\n"
        
        for table_name in inspector.get_table_names():
            for fk in inspector.get_foreign_keys(table_name):
                plantuml_code += f"{table_name} -- {fk['referred_table']}\n"

        plantuml_code += "@enduml"

        pl = PlantUML(url='http://www.plantuml.com/plantuml')
        output = pl.processes(plantuml_code, outfile=f"{database_name}.{erd_format}", format=erd_format)
        
        return {"status": f"ERD generated successfully as {database_name}.{erd_format}", "output": output}
    except Exception as e:
        return {"error": f"ERD generation failed: {e}"}

async def analyze_relationships(table_name: str):
    """Analyze table relationships."""
    # Implementation to follow
    return {"status": f"Analyzing relationships for table {table_name}."}

async def suggest_optimizations(query_patterns: list):
    """Provide optimization suggestions."""
    # Implementation to follow
    return {"status": "Suggesting optimizations."}

async def create_mock_data(table_name: str, record_count: int):
    """Generate test data."""
    # Implementation to follow
    return {"status": f"Creating {record_count} mock records for {table_name}."}