comprehensive_dev_server/infrastructure_monitor.py
import boto3
from google.oauth2 import service_account
from azure.identity import DefaultAzureCredential

# Global variables to store cloud clients
aws_client = None
gcp_client = None
azure_client = None

async def connect_cloud_provider(provider: str, credentials: dict):
    """Connect to cloud APIs."""
    global aws_client, gcp_client, azure_client
    try:
        if provider.lower() == "aws":
            aws_client = boto3.client(
                'sts',
                aws_access_key_id=credentials.get("aws_access_key_id"),
                aws_secret_access_key=credentials.get("aws_secret_access_key"),
                region_name=credentials.get("region_name")
            )
            aws_client.get_caller_identity()
            return {"status": "AWS connection successful."}
        elif provider.lower() == "gcp":
            gcp_client = service_account.Credentials.from_service_account_info(credentials)
            return {"status": "GCP connection successful."}
        elif provider.lower() == "azure":
            azure_client = DefaultAzureCredential()
            return {"status": "Azure connection successful."}
        else:
            return {"error": f"Unsupported cloud provider: {provider}"}
    except Exception as e:
        return {"error": f"Connection failed: {e}"}

async def monitor_resource_usage(resource_types: list):
    """Monitor cloud resources."""
    # Implementation to follow
    return {"status": "Monitoring resource usage."}

async def track_deployment_status(deployment_id: str):
    """Track deployments."""
    # Implementation to follow
    return {"status": f"Tracking deployment {deployment_id}."}

async def analyze_performance_metrics(service_name: str, time_range: str):
    """Performance analysis."""
    # Implementation to follow
    return {"status": f"Analyzing performance for {service_name} in {time_range}."}

async def generate_iac_templates(resources: list, template_type: str):
    """Generate IaC templates."""
    # Implementation to follow
    return {"status": f"Generating {template_type} templates."}

async def get_cost_optimization_recommendations(account_id: str):
    """Cost optimization."""
    # Implementation to follow
    return {"status": f"Getting cost optimization recommendations for {account_id}."}