comprehensive_dev_server/main.py
import asyncio
from mcp.server import Server
from mcp.server.stdio import stdio_server
from mcp.types import (
    Tool,
)
from comprehensive_dev_server.database_explorer import (
    connect_database,
    explore_schema,
    generate_erd,
    analyze_relationships,
    suggest_optimizations,
    create_mock_data,
)
from comprehensive_dev_server.api_doc_generator import (
    scan_codebase,
    extract_endpoints,
    generate_openapi_spec,
    update_documentation,
    validate_api_docs,
    generate_client_sdk,
)
from comprehensive_dev_server.code_quality_analyzer import (
    analyze_code_quality,
    calculate_complexity_metrics,
    scan_security_vulnerabilities,
    identify_performance_issues,
    assess_technical_debt,
    generate_quality_report,
)
from comprehensive_dev_server.infrastructure_monitor import (
    connect_cloud_provider,
    monitor_resource_usage,
    track_deployment_status,
    analyze_performance_metrics,
    generate_iac_templates,
    get_cost_optimization_recommendations,
)
from comprehensive_dev_server.dev_env_manager import (
    orchestrate_containers,
    manage_environment_variables,
    setup_local_ssl,
    run_database_migrations,
    seed_database,
    backup_dev_environment,
    restore_dev_environment,
)

# Initialize MCP server
server = Server(
    "comprehensive-dev-server",
    "A comprehensive server for development and infrastructure management."
)

# Tool Definitions
DATABASE_TOOLS = [
    Tool(name="connect_database", description="Establish database connection", arguments_schema={"type": "object", "properties": {"connection_string": {"type": "string"}, "db_type": {"type": "string"}}, "required": ["connection_string", "db_type"]}),
    Tool(name="explore_schema", description="Return complete schema information", arguments_schema={"type": "object", "properties": {"database_name": {"type": "string"}}, "required": ["database_name"]}),
    Tool(name="generate_erd", description="Generate ERD diagrams", arguments_schema={"type": "object", "properties": {"database_name": {"type": "string"}, "format": {"type": "string"}}, "required": ["database_name", "format"]}),
    Tool(name="analyze_relationships", description="Analyze table relationships", arguments_schema={"type": "object", "properties": {"table_name": {"type": "string"}}, "required": ["table_name"]}),
    Tool(name="suggest_optimizations", description="Provide optimization suggestions", arguments_schema={"type": "object", "properties": {"query_patterns": {"type": "array", "items": {"type": "string"}}}, "required": ["query_patterns"]}),
    Tool(name="create_mock_data", description="Generate test data", arguments_schema={"type": "object", "properties": {"table_name": {"type": "string"}, "record_count": {"type": "integer"}}, "required": ["table_name", "record_count"]}),
]

API_DOC_TOOLS = [
    Tool(name="scan_codebase", description="Scan for API endpoints", arguments_schema={"type": "object", "properties": {"directory_path": {"type": "string"}, "framework_type": {"type": "string"}}, "required": ["directory_path", "framework_type"]}),
    Tool(name="extract_endpoints", description="Extract API endpoint information", arguments_schema={"type": "object", "properties": {"file_paths": {"type": "array", "items": {"type": "string"}}}, "required": ["file_paths"]}),
    Tool(name="generate_openapi_spec", description="Create OpenAPI specification", arguments_schema={"type": "object", "properties": {"endpoints_data": {"type": "object"}}, "required": ["endpoints_data"]}),
    Tool(name="update_documentation", description="Sync docs with code changes", arguments_schema={"type": "object", "properties": {"repo_path": {"type": "string"}}, "required": ["repo_path"]}),
    Tool(name="validate_api_docs", description="Validate API documentation", arguments_schema={"type": "object", "properties": {"spec_file": {"type": "string"}}, "required": ["spec_file"]}),
    Tool(name="generate_client_sdk", description="Generate client SDKs", arguments_schema={"type": "object", "properties": {"openapi_spec": {"type": "string"}, "language": {"type": "string"}}, "required": ["openapi_spec", "language"]}),
]

CODE_QUALITY_TOOLS = [
    Tool(name="analyze_code_quality", description="Run comprehensive analysis", arguments_schema={"type": "object", "properties": {"directory_path": {"type": "string"}, "language": {"type": "string"}}, "required": ["directory_path", "language"]}),
    Tool(name="calculate_complexity_metrics", description="Calculate code complexity", arguments_schema={"type": "object", "properties": {"file_path": {"type": "string"}}, "required": ["file_path"]}),
    Tool(name="scan_security_vulnerabilities", description="Security analysis", arguments_schema={"type": "object", "properties": {"codebase_path": {"type": "string"}}, "required": ["codebase_path"]}),
    Tool(name="identify_performance_issues", description="Performance bottlenecks", arguments_schema={"type": "object", "properties": {"profile_data": {"type": "object"}}, "required": ["profile_data"]}),
    Tool(name="assess_technical_debt", description="Technical debt analysis", arguments_schema={"type": "object", "properties": {"repository_path": {"type": "string"}}, "required": ["repository_path"]}),
    Tool(name="generate_quality_report", description="Generate reports", arguments_schema={"type": "object", "properties": {"analysis_results": {"type": "object"}}, "required": ["analysis_results"]}),
]

INFRA_TOOLS = [
    Tool(name="connect_cloud_provider", description="Connect to cloud APIs", arguments_schema={"type": "object", "properties": {"provider": {"type": "string"}, "credentials": {"type": "object"}}, "required": ["provider", "credentials"]}),
    Tool(name="monitor_resource_usage", description="Monitor cloud resources", arguments_schema={"type": "object", "properties": {"resource_types": {"type": "array", "items": {"type": "string"}}}, "required": ["resource_types"]}),
    Tool(name="track_deployment_status", description="Track deployments", arguments_schema={"type": "object", "properties": {"deployment_id": {"type": "string"}}, "required": ["deployment_id"]}),
    Tool(name="analyze_performance_metrics", description="Performance analysis", arguments_schema={"type": "object", "properties": {"service_name": {"type": "string"}, "time_range": {"type": "string"}}, "required": ["service_name", "time_range"]}),
    Tool(name="generate_iac_templates", description="Generate IaC templates", arguments_schema={"type": "object", "properties": {"resources": {"type": "array", "items": {"type": "object"}}, "template_type": {"type": "string"}}, "required": ["resources", "template_type"]}),
    Tool(name="get_cost_optimization_recommendations", description="Cost optimization", arguments_schema={"type": "object", "properties": {"account_id": {"type": "string"}}, "required": ["account_id"]}),
]

DEV_ENV_TOOLS = [
    Tool(name="orchestrate_containers", description="Manage Docker containers", arguments_schema={"type": "object", "properties": {"compose_file": {"type": "string"}}, "required": ["compose_file"]}),
    Tool(name="manage_environment_variables", description="Environment management", arguments_schema={"type": "object", "properties": {"project_path": {"type": "string"}, "env_vars": {"type": "object"}}, "required": ["project_path", "env_vars"]}),
    Tool(name="setup_local_ssl", description="SSL certificate management", arguments_schema={"type": "object", "properties": {"domain_name": {"type": "string"}}, "required": ["domain_name"]}),
    Tool(name="run_database_migrations", description="Database migrations", arguments_schema={"type": "object", "properties": {"database_config": {"type": "object"}, "migration_path": {"type": "string"}}, "required": ["database_config", "migration_path"]}),
    Tool(name="seed_database", description="Database seeding", arguments_schema={"type": "object", "properties": {"database_config": {"type": "object"}, "seed_data": {"type": "string"}}, "required": ["database_config", "seed_data"]}),
    Tool(name="backup_dev_environment", description="Environment backup", arguments_schema={"type": "object", "properties": {"environment_name": {"type": "string"}}, "required": ["environment_name"]}),
    Tool(name="restore_dev_environment", description="Environment restoration", arguments_schema={"type": "object", "properties": {"backup_path": {"type": "string"}}, "required": ["backup_path"]}),
]

ALL_TOOLS = DATABASE_TOOLS + API_DOC_TOOLS + CODE_QUALITY_TOOLS + INFRA_TOOLS + DEV_ENV_TOOLS

TOOL_HANDLERS = {
    "connect_database": connect_database,
    "explore_schema": explore_schema,
    "generate_erd": generate_erd,
    "analyze_relationships": analyze_relationships,
    "suggest_optimizations": suggest_optimizations,
    "create_mock_data": create_mock_data,
    "scan_codebase": scan_codebase,
    "extract_endpoints": extract_endpoints,
    "generate_openapi_spec": generate_openapi_spec,
    "update_documentation": update_documentation,
    "validate_api_docs": validate_api_docs,
    "generate_client_sdk": generate_client_sdk,
    "analyze_code_quality": analyze_code_quality,
    "calculate_complexity_metrics": calculate_complexity_metrics,
    "scan_security_vulnerabilities": scan_security_vulnerabilities,
    "identify_performance_issues": identify_performance_issues,
    "assess_technical_debt": assess_technical_debt,
    "generate_quality_report": generate_quality_report,
    "connect_cloud_provider": connect_cloud_provider,
    "monitor_resource_usage": monitor_resource_usage,
    "track_deployment_status": track_deployment_status,
    "analyze_performance_metrics": analyze_performance_metrics,
    "generate_iac_templates": generate_iac_templates,
    "get_cost_optimization_recommendations": get_cost_optimization_recommendations,
    "orchestrate_containers": orchestrate_containers,
    "manage_environment_variables": manage_environment_variables,
    "setup_local_ssl": setup_local_ssl,
    "run_database_migrations": run_database_migrations,
    "seed_database": seed_database,
    "backup_dev_environment": backup_dev_environment,
    "restore_dev_environment": restore_dev_environment,
}

@server.list_tools()
async def list_tools():
    """Return a list of all available tools."""
    return ALL_TOOLS

@server.call_tool()
async def call_tool(name: str, arguments: dict):
    """Router for all tool calls."""
    if name in TOOL_HANDLERS:
        return await TOOL_HANDLERS[name](**arguments)
    return {"status": f"Tool '{name}' not implemented yet"}


if __name__ == "__main__":
    stdio_server(server)